fix: Melhorias, pequenas correções e adicionado README.md

+++ Utils/ValidationHelper.cs
@@ -4,7 +4,7 @@ namespace Livraria.Utils
     {
         public static bool IsNumeric(string value)
         {
-            return !string.IsNullOrWhiteSpace(value) && int.TryParse(value, out _);
+            return !string.IsNullOrWhiteSpace(value) && double.TryParse(value, out _);
         }
 
         public static bool IsString(string value)
